test_uniform_block_new
